rename singleton -> global_class; rename singleton.py -> decorators.py
