+++ Code/systems/decorators.py
@@ -1,0 +1,27 @@
+def global_class(cls):
+    """Декоратор для обеспечения того, что класс следует паттерну Singleton, то есть создается и используется только один экземпляр класса во всем приложении.
+
+    Args:
+        cls (type): Класс, к которому применяется декоратор.
+
+    Returns:
+        function: Функция, которая возвращает единственный экземпляр класса.
+    """
+    instances = {}
+
+    def get_instance(*args, **kwargs):
+        """Возвращает единственный экземпляр класса, создавая его, если он еще не существует.
+
+        Args:
+            *args: Позиционные аргументы для передачи в конструктор класса.
+            **kwargs: Именованные аргументы для передачи в конструктор класса.
+
+        Returns:
+            object: Единственный экземпляр класса.
+        """
+        if cls not in instances:
+            instances[cls] = cls(*args, **kwargs)
+        
+        return instances[cls]
+
+    return get_instance
